feat(kernel): add a barebone code for describing cpu architecture as a TREE
For now, we will use the default one generated (cpu 0 is near cpu 1, cpu 2 with 3...), because QEMU don't generate an SRAT table, making us unable to generate a tree for now.

This tree will be used by the scheduler when deciding if a task can be used for/by a processor.